assets_management_service/requirements.txt
asgiref==3.7.2
class-registry==2.1.2
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
filters==1.3.2
filters-django==1.0.5
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn==20.1.0
cryptography==39.0.1
